權限設置, 調整post功能

+++ routes/api.php
@@ -2,6 +2,7 @@
 
 use App\Http\Controllers\PostController;
 use App\Http\Controllers\AuthController;
+use App\Http\Controllers\UserGroupController;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
 
@@ -23,10 +24,12 @@ Route::get('posts', [PostController::class, 'index']);
 Route::get('posts/{id}', [PostController::class, 'show']);
 
 // protect
-Route::group(['middleware' => 'auth:sanctum'], function(){
-    Route::post('posts', [PostController::class, 'store']);
-    Route::put('posts/{id}', [PostController::class, 'update']);
-    Route::delete('posts/{id }', [PostController::class, 'destroy']);
+Route::group(['middleware' => 'auth:sanctum'], function(){ // 無登入使用者不可使用以下api
+    Route::post('setGroup', [UserGroupController::class, 'setGroup'])->middleware(['ability:admin']);
 
+    Route::post('posts', [PostController::class, 'store'])->middleware(['ability:admin,manager']);
+    Route::put('posts/{id}', [PostController::class, 'update'])->middleware(['ability:admin,manager']);
+    Route::delete('posts/{id}', [PostController::class, 'destroy'])->middleware(['ability:admin,manager']);
     Route::post('logout', [AuthController::class, 'logout']);
 });
+
